Régler les problèmes de compilation fonction.h
Nous avions perdues le code de fonction.h/fonction.cpp, j'ai pu
retrouver le code et faire les modifications nécessaires pour que
prendre en compte les changements fait lors de la rédaction de l'énoncé.
Nous avons encore des corrections à apporté.

+++ laboratoires/lab02-oop1/solution/fonctions.h
@@ -8,7 +8,7 @@ void afficherMenu(Menu &);
 int demanderChoix(Menu &menu);
 void afficherListe(Personne &);
 void ajouterTache(Personne &);
-int demanderNumeroTache(std::string, Personne &);
+size_t demanderNumeroTache(std::string, Personne &);
 void marquerFaite(Personne &);
 void echangerTaches(Personne &);
 void sauvegarderListe(Personne &);
